[assembly] Core product assembly config is now testonly
The core.gni file is not, but the product assembly config
is testonly, which allows it to contain testonly items.

This required forwarding testonly through to a number of
other templates to properly mark them all as testonly
when the product assemblyy is testonly.

Change-Id: I88f60a2da02c398f9f29ab046e531e3cc9353802
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729323
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>