[fdf] Create DFv2-specific core bot tests
Create buildbot groups `//bundles/buildbot:core_dfv2` and
`//bundles/buildbot:host-tests_dfv2` which are similar to their
counterparts, `core` and `host-tests`, except they do not contain tests
that fail when DFv2 is enabled and using chromebook.

These groups are to be used by the core.x64-dfv2-release and
core.chromebook-x64-dfv2-release builder bots in a future CL so that the
builders can pass and be monitored by the Fuchsia build gardeners.  This
is in order to prevent regressions in DFv2.

Bug: 102234
Change-Id: I17497c37be86d525ca2daf4d9579348e334d1f14
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719645
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>