[ffx] Add back `ffx component run-legacy` plugin
The plugin `run-legacy` is restored exactly as-is, with no
modifications.

This plugin is used by Chromium folks for running v1 components.
It is currently causing the SDK rolls into Chromium to fail
due to Chromium's dependency on `ffx component run-legacy`.

The dependency was not caught at the time because the LSC presubmit for
Chromium was broken when commit 50e4938648c28379e227aa7b17d0776ef7e61873
was merged.

Change-Id: Ic3d1a88a947ba9167439aed1bb7948244b5c2249
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729607
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>