[nxpfmac] Fix flaky dataplane tests
There was a small chance that dataplane tests would flake during
destruction because a raw pointer access. If the pointer was already
deleted the destructor would access freed memory.

With recent changes to the DataPlane destructor this complicated setup
can be simplified. Now we just need to wait for the async remove watcher
thread to complete after deleting the dataplane. Also use a shared
pointer to the device object to make sure it's kept alive for as long as
needed.

Test: Repeated unit test runs with asan enabled
Bug: 110221
Change-Id: I824f552850c9182f79cfdad4990eb583bc01780a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729624
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>