[ffx emu] Use `cache=unsafe` for drives to reduce the physical disk writes
The definition and choices of `cache` parameter can be found at
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/virtualization_tuning_and_optimization_guide/sect-virtualization_tuning_optimization_guide-blockio-caching
(There must be an official one in qemu.org, but I cannot find it.)
In short, `cache=unsafe` disables the flush command to write-back the
caches in both host and guest to the physical disk. The data may be
lost if the emulator shuts down unexpectedly. But in practice,
there is no reason to preserve data in the qemu instance and
writing-back is a waste of the SSD lifespan.

Change-Id: Ia4f6eb467ecf5121e481ba38adde96f8ce5fa91c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729186
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Commit-Queue: Zijie He <zijiehe@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>