[virtualization] Add Stop support to the guest tool
This adds a new stop command to the guest CLI tool. A user can use
`guest stop <guest_type>` to send a graceful shutdown command to the
guest, or force a stop by adding -f.

This reuses the existing virtio-consoles to send commands, which isn't
the ideal scenario since the console may be doing something else. We can
change this once (if) we move to supporting multiple consoles.

Bug: 104989
Tested:
1) Launched Debian
2) Ran `guest stop debian`
3) Confirmed debian received the graceful stop command and stopped
4) Restarted debian (without restarting host, etc.)
5) Ran `guest stop debian -f` and confirmed that debian stopped abruptly
6) Repeated for Zircon (Termina can shut down but not restart due to
   termina guest manager state)

Change-Id: I3bfe381eb5d4975658d8abab1b9dc9c107e3a461
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727924
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>