[build][bazel] Add //build/bazel:generate_fuchsia_sdk_repository
Add a new GN target that ensures that the @fuchsia_sdk repository
is properly generated, and creates a symlink to it with a stable path
if $BUILD_DIR/gen/build/bazel/fuchsia_sdk.

This is necessary due to the fact that the exact repository path is
an implementation detail of Bazel and can vary based on a multiple
factors (i.e. whether BzlMod is enabled, and how repositories are
generated using extensions).

Bug: 106543
Change-Id: I0c57a83a0722fb52198809edc52bd1e9cff30b1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730783
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>