[component][fuzzing] Add retry to fuzz tests
stderr from spawned processes is sometimes truncated, most likely as a
result of the socket draining task being dropped on task termination as
described in fxbug.dev/95602.

This results in certain workflows encountering errors due to missing
stderr from libFuzzer. In particular, this affects "brief" workflows
like `TryOne` which is used by `Engine::RunTests`. The rate of incidence
is low and in the single digit percentage at most, and can be made
miniscule by retrying a fixed number of times in the event of this
specific error.

Bug: 109100
Change-Id: I1ce5700da744648de83b3b3c5542f3d149914717
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730863
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>