[driver_manager] Pass banjo symbol from DFv1 to DFv2
Support banjo in the DFv2 alongside implementation.
This is done by passing the symbol from a DFv1 driver
out to DriverManager, and filling it in to DFv2 drivers that
are colocated.

The lifecycle test is updated to call it's parent's banjo
function.

TEST: fx test dfv2_v1_lifecycle
Change-Id: I2c669664386a70a84802b24c4d2cc30c5afeb966
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727344
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Christopher Anderson <cja@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>