[component_manager] Fix stress test total duration
The test duration was set to the string value "60 * 60" in the
test manifest.

The value is expected to be parsed as an integer, not as math
that needs to be evaluated. So it defaulted to running for 6 hours.

Reduced the time to 1800 seconds (30 minutes) because the logs
produced from this test were very large and symbolization would
hang, failing the stress test builder.

Change-Id: Id41568fe0055935d14e0eea1268fd2e9b65ec693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/730844
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>