[fxfs] Unlock volumes read-only for fsck
Per-volume fsck was not re-locking the volume after checking it, which
is problematic for two reasons: (a) it's a possible security issue, and
(b), it prevents the volue from being mounted (which prevented this from
actually being used in a real environment anyways).

However, we couldn't just re-lock the volume, since fsck holds write
locks on the entire filesystem and re-locking triggers flushes.

Fix this by having fsck unlock the volume as read-only, which prevents
all flushes from occurring on the volume.  The encrypted mutations in
the journal are buffered in the lock state of the store while it's
unlocked read-only.  fxbug.dev/110284 describes a more graceful way we
can deal with this later.

Bug: 106845
Change-Id: Ia4e43c95868f5be6d509b6ef49e664ebe1d8ac5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729483
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>