[ui][input-pipeline] gestures: rename details
Rename `MismatchData` and related types. This is in
preparation for using the types to carry explanations
of why a gesture ended.

Although we could call the end of a gesture a "mismatch"
as well, doing so would likely be confusing: readers
could reasonably assume that a mismatch is related to
the `Matching` state.

Bug: 107948
Test: fx test input_pipeline_lib_tests
Change-Id: Ibe5eb8cf64410aea8fb486564378832c5754aca3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728823
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>