[nxpfmac] Don't complete destroy interface early
The wlanphy DestroyIface call would call DdkAsyncRemove on an interface
and then complete the request. However the interface is not completely
torn down until DdkRelease is called on it and the interface destructor
completes. This caused problems when a fast sequence of DestroyIface and
CreateIface occurred. The interface was not completely destroyed when
the create call arrived. Even worse is that the CreateIface call would
succeed but silently fail to create the network device port, thus
preventing any data from being received or transmitted on the interface
will still appearing to be successfully created.

Fix this by ensuring that DestroyIface does not complete the request
until the interface is fulle destroyed.

Test: ACTS test_start_stop_client_connections, manual tests
Change-Id: I2f97ad2a58e75dad7218013e3062cdfd9af7fc45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729184
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>