[nxpfmac] Only complete DdkInit once FW is ready
DdkInit would synchronously reply to the init call even if firmware
initialization was in a pending state. In situations where the wlanstack
is already up and running when this happens this meant that the
wlanstack would start making calls into the driver that would attempt to
use mlan before it was ready, leading to crashes as mlan tried to
dereference null pointers.

Fix this by detecting when firmware initialization is still pending and
deferring the transaction reply until the initialization completes.

Test: Induce a crash in the driver that causes a rebind
Change-Id: Ic9ec528b200422af305d81c539d510c95cbf2c0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729185
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>