[pbms] Structured UI for product-bundle, take 2
This is a re-working of fxrev.dev/665178. In that CL, the structured UI
module was introduced to the PBMS library for FFX, which provided much
cleaner, more user-friendly output for calls to `ffx product-bundle`.

This CL takes the original change and merges it with the current
code-base, including some minor changes to account for PBv2 code, etc,
and also addressing the outstanding comments for that previous CL.

Bug: 96950
Bug: 109933

Change-Id: Ib08dde7e7b90edf76438d3b9e758b7ab5d554a70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/729490
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>