[vaapi][fuzz] Allocate stub surfaces
The TiledSurfaceManager expects surfaces to be unique so we have to
model that behavior in our vaStubs. vaCreateSurfaces() will now return
unique surface ids and vaDestroySurfaces() will return those surfaces
back to the pool of available surfaces.

Also switch a ZX_DEBUG_ASSERT to ZX_ASSERT. Doesn't make sense that
the uniqueness check is a debug assert while the removal is a normal
assert.

Bug: 103532
Change-Id: I1ff822309a4be067931030517cac29c724fa6b56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731708
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>
Reviewed-by: John Bauman <jbauman@google.com>