[pbms] Add product-bundle remove command
This CL adds a new subcommand, `ffx product-bundle remove`. This command
takes an optional product bundle name/url, removes that product bundle
from the cache, and unregisters the repository for that product bundle.
If no product bundle is specified, it uses the same logic as other
product bundle tools to find the best match. The user is prompted for
confirmation before anything is removed, unless they include the
--force/-f flag in the command. There is also a --all/-a flag which will
remove all product bundles in the user's cache.

To support finding product bundles that all already present on the host,
the `select_product_bundle()` code is adjusted to allow the caller to
request only locally cached bundles when resolving the user's input. The
`list` and `get` subcommands still operate as before.

Bug: 99227

Change-Id: I270807c51fca1903d9486280047c013207d8f96c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728467
Reviewed-by: Megan Batty <mgnb@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>