[overnet][hoist] Work around short SUN_LEN issues
This makes it so that when we try to use a long path to a socket, we
temporarily create a short path 'landing link' in TEMP to bind or
connect to it through. For the most part, we still talk only about the
real socket path and only use it for the connecting/binding calls, so
that it doesn't interfere much with other things.

Note that it actually creates a symlink to the directory the socket is
in, not directly to the socket. This makes sure that we behave properly
in terms of creating the socket file and probably makes it more likely
this will eventually work correctly on windows (though that will likely
come with other challenges, and the likelihood is still low).

While I was in here I cleaned up some of the dependency logic and
simplified the gn template and use statements a bit.

Fixed: 109815
Change-Id: Ia127f0daa7bad757cb16deae0d9a1a96427bb32e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731737
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>