[wlan][hw-sim] check security in connection status
Previously, when tests checked the status of connection through WLAN
Policy client state updates, only SSID was checked. After this check,
the security type is checked as well.

In particular, this would help in a flake from fxbug.dev/108037 where an
auto connect could result in a connection succeeding as a network is
removed. In that case the 'connected' update would be counted toward a
following test, even if the security type used in the next test is
different. Checking the security type would make that test fail sooner
so the failure would be easier to triage.

Bug: none
Test: ran hw-sim tests
      (fx test//src/connectivity/wlan/testing/hw-sim/test)
Change-Id: I51068c593904ee3dda8aa2c524f5109c6ee53260
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/728628
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>