[virtualization] Add termina stop RPC to guest tool
This provides a way to gracefully stop Termina. Please note that we're
not emulating I8042 correctly preventing us from cleanly shutting down
when the reboot syscall is used.

We crash and stop the VM however, so mission (somewhat) accomplished.

Bug: 104989

Tested=Started and stopped termina on a NUC

Change-Id: I504faf0d203b49df75c9d98591282b24b6cd4fb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731586
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>