[activity] Remove the implementation of fuchsia.ui.activity.Tracker in the legacy activity service.
The protocol API is not removed as part of this change since that
requires going through the LSC process, and will be done so separately.

This change does not remove the implementation of other protocols served
by the legacy activity service, namely fuchsia.ui.activity.Provider and
fuchsia.ui.activity.control.Control. This is because existing users rely
on this functionality, and the service cannot be removed entirely until
they are migrated off of them.

Bug: 103843

Change-Id: I36d629d064104dae1c3f93a52f204746bc4c6b30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714982
Commit-Queue: Caroline Liu <carolineliu@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Gary Bressler <geb@google.com>