[bt][fastpair] Define Fast Pair as a static child of bt-init
The Fast Pair Provider component will not exist on all products. It
both uses and exposes the `sys.Pairing` capability.

Define the Fast Pair Provider component as a static child of `bt-init`
and add the relevant capability routes.

Update bt-init to conditionally route `sys.Pairing` requests to Fast
Pair if the component exists. This logic is identical to that used in
bt-rfcomm & `bredr.Profile`.

Downgrade the log message emitted when the child component cannot be
resolved. This will likely occur fairly often when RFCOMM/FastPair
functionality is not supported on the device.

Update integration tests that define a test bt-init realm to point to
the Fast Pair Provider component. However, because Fast Pair
functionality is not relevant to this tests, omit it from the test
package so that the URL doesn't resolve. Any URL resolving warnings can
be safely ignored.

Bug: 108241
Test: fx test bt-init-smoke-test
Test: fx test bt-access-integration-tests (Any integration test that
relies on the bt-test-harness will exercise this new change)
Test: fx test bt-init-unittests

Change-Id: I6aba1f06c9231598c25dc109603ff2b790ee203e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718322
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>