[system-update-checker] Update tests per RFC6761
RFC6761 defines special-use domain names, including domains to be used
in documentation. 'system-update-checker' unit tests encode fuchsia.com,
which is a name that may exist n DNS. It can be clear whether reliance
on the function of that domain within DNS is relevant to the test.

By parameterizing the function that uses the fuchsia.com url and
creating a separate `.test` URL for test update urls, it ensures that
the tests will remain independent and not reliant on the `com` domain
within DNS.

Note: created a shared variable in main.rs that is used within apply.rs
and check.rs, per reviewer comments

Test: system-update-checker

Fixed: 109814
Change-Id: Ib4bc5368841cedae53e3304a0925b0fc35c777b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727976
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>