[power][test] Allow disabling temperature filter
Adds support to disable temperature filtering in integration tests. The
switch is added to Power Manager's structured config, and is always
false by default.

It's useful to disable temperature filtering in tests because the filter
is dependent on time which is not predictable in a testing environment.

Bug: 104266
Test: fx test //src/power/power-manager
Change-Id: I97958833675fb18b0bc6d730575d56161faf7cb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727970
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Patrick Shickel <pshickel@google.com>
Reviewed-by: Didi She <didis@google.com>