[audio] Split Thread into `fidl/` and `mix/` parts
Graph nodes are split into a class that lives in `fidl/` (Node) and a
class that lives in `mix/` (PipelineStage). The Node objects are updated
eagerly from the FIDL thread while the PipelineStage objects are updated
asynchronously.

This CL adds a similar structure to Thread. Now we have GraphThread (in
`fidl/`) and PipelineThread (in `mix/`).

Splitting threads in this way makes some graph operations simpler. For
example, Graph::DeleteThread needs to know if there are any consumers
still using that thread. It can't ask the PipelineThread because that
object is updated asynchronously. The simplest approach is to manage
that accounting in both `fidl/` and `mix/`.

The main changes are in the following files (other changes are mostly
mechanical):

mix/pipeline*thread*
fidl/graph*thread*
fidl/testing/fake_graph*

Bug: 87651
Change-Id: I5ed5c5e3f26f1462df4174ad64e6f9c14c12732b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731302
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>