[netstack3] Drop ctx arg from timer trigger fns
...which is currently just a pass-through to the timer handler
callback. A separate helper is used to bind arguments to the timer
handler method.

This is done to enable a single variant of `trigger_XXX` methods
on the `DummyTimerCtxExt` trait across the different types of
references (`&mut T` and `&T`).

Note that the ctx arg is only removed from methods that may trigger
multiple timers because there is no way to have a type that is generic
over both immutable and mutable references. Such a generic type needs to
support "re-borrows" so that we can do things like pass a mutable
reference (behind an opaque type) to a callback function in a loop but
this does not seem to be supported today so the solution was to remove
the pass-through ctx and bind it to the callback at the call-sites.
`trigger_next_timer` keeps the pass-through ctx arg because it may only
trigger at most 1 timer so that method does not need to support
"re-borrows".

Bug: 48578
Change-Id: I81b5b6a24bbef1b88c238c801fa0234bd52f41e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726206
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>