[netstack] Avoid leaking space in the routes slice
When a route is removed from the routes slice, we reused the existing
slice's allocation rather than allocating a new one. This means that
routes[len(routes):cap(routes)] could potentially be leaked if we never
overwrite that memory with new route entries.

To fix this, zero out "dead" route entries at the tail of the routes
slice whenever we remove a route.

In a related issue, because we were reusing the existing routes slice
when removing routes, the route slice never shrinks in capacity -- only
grows. Fix this by copying all routes into a new slice when more than
half the capacity is unused.

Test: netstack-gotests
Change-Id: Ie0c0fba4c8d8fcf2211e52ce90a06d5b64b23657
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731967
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>