[ui] include fuchsia.media.sounds.Player in the minimal ui realm
Scene Manager (namely, Input Pipeline) will use this protocol to play a sound from the FactoryResetHandler. I made this optional so that it can be available in the minimal realm (where it is needed), and omitted from the regular ui realm because it hasn't been indicated as needed before.

Change-Id: I8ebb1942e1c8474f7255d33c1c87ef466023669a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732296
Reviewed-by: Alexander Brusher <abrusher@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>