[radarutil][test] Wait for a VMO to be unlocked instead of failing
Previously the fake radar driver in this test would return an error if
no VMOs were unlocked when a new burst was ready. It also used a burst
period of 33ms in order to match a real radar sensor, which mostly
masked the VMO locking issue. Flakes were still possible however, so
this change makes the fake driver wait for a VMO to be unlocked instead
of reporting an error. It also makes the fake driver report bursts as
fast as possible instead of waiting 33ms, which makes the test take much
less time overall.

Bug: 110572
Multiply: radarutil-test: 100
Change-Id: I289c8a0de53f6feaefa0a35fd3de954beaa1ffdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/731936
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>