[update] Update docs per RFC6761
RFC6761 defines special-use domain names, including domains to be used
in documentation. This is a nit fix to use fuchsia.example in comments
instead of the fuchsia.com, and is part of an inititaive to replace
these domains across parent directories to comply with the special-use
domain names and avoid potential DNS issues.

Fixed: 109838
Change-Id: Ife143f4f0927c31f8888c1f2551c598b69c1269f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727365
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>