[overnet] Cleanup stream ownership in circuit router
Just some API rearranging. This makes the user responsible for
allocating most of our streams and channels, which is more flexible for
the user. We also shed some abstraction around some of the channels
without much loss of clarity.

The only other major change is connect_node now allows you to not pass a
control stream, in which case it will use the first incoming stream as
the control stream.

Change-Id: I11558177b9addd64b4a4e96d330dbd9c592f76dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716804
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>