[build] Fix visibility usage in structured config values.
When restricting visibility of a config value file,
GN complains if the target itself isn't added to the
visibility list. This is caused by the Fuchsia target
being depended upon by the group that's used by non-Fuchsia
toolchains, so we need to list them too. However we don't
have variables defined for all the toolchains that might be
involved, so we add a blanket visibility for the target name.

Fixed: 110496
Change-Id: I0234aa5316fb68c658493baaeb4642e464aaf017
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732542
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>