[testsharder] Skip unaffected if affected tests found even if not run.
The build might detect affected tests that don't have an environment
attached and will thus not be run as part of that builder. This will
cause AffectedModifiers() to return 0 modifiers, but the original
affectedTests file did have tests that were detected to be affected. So
we should check that file to determine whether we need to run unaffected
hermetic tests instead of checking for an empty set of affected
modifiers.

Change-Id: Ia15cdd87ceae8ad17d7c82000a195b01231cc990
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733133
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>