[netstack3] Reduce locks held on dev state access
This change wraps all device states around a reference counted pointer
and only holds the device map's read lock until the pointer is cloned.
Once the pointer is cloned, we no longer have a need to hold the device
map's lock since the device specific state is protected by its own
locks.

This prepares for a later change which requires accessing a device state
while in a `with_XXX_state` callback method which holds some lock(s)
protecting device state. If we continued to hold the devices map lock,
accessing the device's state while in the `with_XXX_state` results
in a panic because we do not allow taking the devices map lock
recursively so we need to release the devices map lock as soon as we
can.

Note a later change will hold device IDs are device references so this
will no longer be a concern at that time because we do not need to
access the devices map when we already have the device reference.

Bug: 48578

Change-Id: I401d6e25df957a5883cbddee793f92b56c4f1adc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733225
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>