[iwlwifi] Remove banjo dependencies
The iwlwifi driver uses some of the banjo generated definitions and
since banjo will not be available oot, we are copying over these definitions
locally to allow the driver to compile once it is oot.

Note, since these definitions are no longer coupled to their respective
.fidl files, the driver needs to ensure that it always translates values
between the fidl interface and the local ones, and not copy them as is.
Doing so could lead to bugs when they get out of sync.

Bug: 99490
Test: Existing tests cover this change

Change-Id: I1f5f4c1fe2dd17e043129d8404f8f96257ffe0e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676148
Reviewed-by: Louis Lo <yjlou@google.com>
Commit-Queue: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>