[audio] CreateEdge and DeleteEdge move subtrees across threads
When created an edge A -> B, we currently move A to B's thread.
Similarly, when deleting edge A -> B, we currently move A to the
detached thread.

However, we actually need to move an entire subtree of nodes above A.
This is implemented with a new function in reachability.h,
MoveNodeToThread, which operates on a new property, Node::is_consumer,
as described in the comments in reachability.h.

Bug: 87651
Change-Id: Ifcf5fe433d92a2948d116e1e3f008726c73b2236
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732567
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>