[gazelle] Add `wm-tests` to CQ.
I had failed to restrict that test to devices with graphics support. I
got a protip that `environments = vulkan_envs` is the way to do that,
so I've added that to `wm-tests` and to the AppKit tests as well, to
make it clearer what the `AEMU` reference is for.

Change-Id: I425024991a006e597aa16e12d556ed515fe26f97
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/731782
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>