[netstack3] Implement RX device queue for loopback
...so that loopback packets are never handled synchronously with any
send path. That is, this change decouples the TX and RX paths for
loopback, resulting in any call into core only ever sending a packet
or receiving a packet, but never doing both. This allows (in later
changes) for packets to be sent while holding a socket lock without
a chance of a deadlock when the sent packet is expected to be
delivered to the sending socket itself.

Bug: 48578

Change-Id: Ib4df6d9226dcc4df7ced60eafcd082bc878e8d9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733043
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>