[intel-i915-tgl] Only write transcoder disable config when transcoder is enabled.
On Tiger Lake, if the driver sets `enabled_target` to false and writes
the TRANS_CONF register when the transcoder is disabled, the whole
display hardware will crash and cause the system to reboot.
Instead we must only write the register when the register is currently
enabled (`enabled` bit is 1).

Bug: 109368
Change-Id: I8b840af40c2375306b51c50076aba2ee545d5ef4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/734044
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>