[driver_manager] Replace Device::IteratorFactory
Clean up the code to get the Device's children.
Instead of using a hand-rolled iterator and iterator-factory
class, we can get the children ahead of time and stick them
in an STL container so we don't have to write the iterator code.

This should have identical behavior to the orginal in terms of
which children are present.

The main benefit is cleaning up the logic and the header file.

Change-Id: Ie4e7f3c83bcae3c55f607fe6ebe62811c8960d72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733624
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>