[testsharder] Don't set StopRepeatingAfterSecs if user specified runs.
Before, we were setting it to the target duration, but that would cause
the shard to exit early before running all specified runs and there
would be no signal that it ran fewer tests than specified. See the
multiplied shard for go/bbid/8801575033527660129 which only ran 49 runs
when the user specified 300 in the commit message. This could be
confusing, so we should just try to run all runs, and if it ends up
taking too long, the shard will just time out which will at least
provide signal that we couldn't run it for the full count (so the user
will need to specify a lower count).

Multiply: web-semantics-test: 60
Bug: 76711
Change-Id: If55663e92586832030e4add0f1ac8a1c8e0aacb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/733845
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>